Moving constants to each provider
